moved footlightfiendtest outside of merfolktrickstertest
