+++ Mage.Tests/src/test/java/org/mage/test/cards/continuous/MerfolkTricksterTest.java
@@ -106,27 +106,6 @@ public class MerfolkTricksterTest extends CardTestPlayerBase {
         //assertAllCommandsUsed(); // uncommenting this will force a failure since PlayerA cannot do a command to target Trickster, as expected
     }
     
-    // Making sure Footlight Fiend working correctly here first...    
-    @Test
-    public void test_GrizzlyBearBlocksFootlightFiend_BothDie()
-    {
-        addCard(Zone.BATTLEFIELD, playerA, "Footlight Fiend"); // (R/B) 1/1 on death pings any target for 1
-        addCard(Zone.BATTLEFIELD, playerB, "Grizzly Bears"); // (G) 2/2
-        
-        attack(1, playerA, "Footlight Fiend");
-        block(1, playerB, "Grizzly Bears", "Footlight Fiend");
-        addTarget(playerA, "Grizzly Bears");
-        
-        setStopAt(1, PhaseStep.END_COMBAT);
-        execute();
-        
-        assertLife(playerA, 20);
-        assertLife(playerB, 20);
-        assertGraveyardCount(playerA, "Footlight Fiend", 1);
-        assertGraveyardCount(playerB, "Grizzly Bears", 1);
-        assertAllCommandsUsed();
-    }
-    
     @Test
     public void test_TricksterBlocksTibaltToken_Survives()
     {
